requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
httpx
